refactor(util/ast) - util for finding closest function symbol to given location
